Queue Length stats added to MetricsDashboard
1) The internal queue structure in inet DropTailQueue model (which
originally was a cQueue class) is replaced by cPacketQueue in order to
be able to drop the queue length stats out of queue not in terms number
of packets but also in terms of bytes.

2) The queuelen stats signal is emitted only before a packet is added to
the queue. Originally this signal was emitted both before a pkt was
added to queue and after packet was dequeued. This way of emitting the
signal is useful only for finding the time avg of queueulen.

3) HomaTransport.cc is also modified to drop queuelen stats (number of
byte not yet sent when a new message is receiver from application for
transmission).

4) Following stats are added to the MetricsDashboard:
    1- Traffic Characteristics (rates, bytes and dutycycles) at
    different points in the topology

    2- Queue Lengths in number of bytes and pkt at different queues in
    the topology