Correct queueLen empty stats and one pkt stats
previously, the queue len stats was a being calculated based on the
queue.len which turned out not to be correct and doesn't account for the
packet that in currently being serialized on the wire by the MAC so the
stats for queueEmpty and queueLenOne wasn't correct. Now, these two
stats are collected by comparing the wait time a packet in the queue and
the serialization time of the currently transmitting packet when the new
packet is added to the queue.