Implements TrafficPacer
COMILED NOT TESTED
The ByteBucket changes name to TrafficPacer in this commit. The
TrafficPacer not only regulates the grant transmission times but also
will take into account the number outstanding bytes not yet received
and alwasy tries to cap this number of bytes at a max value (ie. one
BandWidth Delay Product).