Fix transport queue length stats
After adding the unsched pkts to transport, a new message send to
transport can completely transmit without any grant required through the
unscheduled packets. So after sending a req. and unsched. packets we
should check if the message is completely sent and remove it from the
list of outbound messages.