Implements priority classifier for queues
Impelemted the packet classifier for network queues when
PseudoIdealPriorityTransport is used as the transport. This function
HomaPkt::compareSizeAndPrios() classifies a packet based on the size of
the message the packet belong to and the priority of the packet
(remaining bytes of message after this packet). This commit also
modifies the DropTailQueue model so that we can choose different
classifiers for the queue depending on if the transport is the
PseudoIdealPriorityTransport or HomaTransport.