A new implementation for cut-through
This implementation is based on two factors:

1. In all of the switches in the network, we set deliverOnReceptionStart
flag of the input gate of MAC modules to true but we set it to false in
the hosts.

2. The new Hook module called Lagger has been added in the input and
output path in each NetworkInterface that in input path annotate packets
with the link speed of the input gate and in the output path adds the
right amount of the delay to the packet if packet is going to an output
link which has a higher speed than the input link.