Implements a new HeavyMiddle fabricated workload
So far, among the different workload that we've had, the
FacebookKeyValue is extremly light tail and the rest of the workloads
are heavy tail. Our set of workloads lacked a workload that is neither
light tail nor heavy tailed. So we have fabricated a new workload we
call Fabricated_Heavy_Middle to address this issue.