New module added to compute avgUnschedRate at RX
Implemented a module in the receiver that tracks the received unsched
packets and compute the avg unsched rate.
The computed rate is used in TrafficPacer to scale the
maxAllowedInFlightBytes down by
"1-fractionOfDataRateThatComesUnscheduled". Whether or not this scale
down is used in the simultion can be controled by the config parameter
"useUnschRateInScheduler".