Fixes a CBF calculation bug in plot script
For computing CBF in SizeDistributionPlotScript.r, we was not taking
into account the extra header bytes that goes onto the wire when sending
a packet. This was a bug that had very little effect on the computed cbf
of heavy tail workload but could have significant effect on the light
tail workloads.