Added infrastructure for unit testing with gtest
The idea is that we use opp_test to create a simple test module. All
this test module should do, it to initialize an empty omnetpp network
and call the gtest main function in the its activity module. From this
point on, gtest will be initialized and performs the unit tests.

The code in the current state compiles, creates the network and
intializes the network but gtest doesn't initialize properly.