Simulated SRBT policy for priority assignment
A new priority assignment policy called Simulated SRBT is added to the
simulator. This policy assigns priorities based on the remaining bytes
of the messages. If Unsched Limit is one RTT worth of bytes that is
probably 6 packets, then at sender the first unsched pkt (ie. req.) gets
priority 5 and last unsched packet get priority 0. At the receiver, tha
last RTT remaining grants for a messsage will be assigned priority
similar to what sender assigns to unsched and the if the remaining bytes
are larger than RTT, then the priority will assigned based on the
adaptive protocol.