Adding a new simple module: RpcHandler
A new module called RpcHandler.ned is added between the applications
(ie .TrafficGenerators) and transport. This module takes all request (ie.
AppMessages) from all applications and translate them to Rpc that will
be given to transport for the transmission. This new module will be the
single place to handle Rpc requests/reponses in the future. This module
is also responsible for statistics collection for messages, a task that
was previously done by the applications.
Moreover, a channel is created between applications and Oracle that will
send a copy of each transmission message to the Oracle scheduler when
the endhost transport is set to MinimalTransport. This will provide the
possiblity for the Oracle to learn about a message in advance of its
transmission time.