Fixes bugs in the compare func. of priority queues
std::priority_queue apperntly is not guaranteed to keep fifo ordering
when it's compare function returns false. This cause a bug for usages of
priority_queue when prio of two element are equal. To prevent this bug
we must make sure the elements are strictly ordered even when prio
numbers are equal.