change pattern of test messages for PRESET_IN_FILE
This commit contains code for new pattern designed for
HostidSizeInterarrival.txt. This file is intended for testing transport
algorithm which specifies a predefined set of message that are transmitted from
predefined set of senders at definitive arrival times. In this commit
the pattern of this file is changed to include a predefined receiver as
well and the code has been added to handle this new pattern.