script to find app goodput for a given load factor
The real avg load is different from application level avg load.
This difference is as a result of header overhead bytes added to the
data packets and control packets. For a desired network load and message
size distribution, this script calculates the application load factor
and average message size on the network.