Fixed the bug causing large 99% slowdown
number of connections per tcp pair was 1 by default. This would cause
the pairs to create connections lazily when the there is no unused
tcp connection between two nodes. It turns out that connection setup
cost would impact the 99% slowdown. We increased this value to 10 which
cause each node to initially create 10 tcp connections to any other
node. This cause 99% slowdown to not have connection setup time included
in them.