Change makefiles to fix heap-overflow bug in tests
Test cases were all failing because of different parts of the test code
were using two differnt compiled versions of HomaTransport module which
was causing a heap-buffer-overflow. This was as a result of makefiles
compiling two different versions of the HomaTransport.cc files. This is
now fixed and only one version of object file is compiled for
HomaTransport.